More fixes for 64 bits arch. Removed some redundant virtual functions (CacheableScriptFile::Seek, GetPosition, CFileText::seek, GetPosition).

+++ src/common/CResourceBase.h
@@ -367,7 +367,7 @@ public:
 
 	bool IsLinked() const;	// been loaded from the scripts ?
 	CResourceScript * GetLinkFile() const;
-	int GetLinkOffset() const;
+	size_t GetLinkOffset() const;
 	void SetLink( CResourceScript * pScript );
     void CopyTransfer( CResourceLink * pLink );
 	void ScanSection( RES_TYPE restype );
